# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ContextResponse" : {
        "properties" : {
          "messageTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "messages" : {
            "items" : {
              "$ref" : "#/components/schemas/MessageResponse"
            },
            "type" : "array"
          },
          "messagesStartSeq" : {
            "format" : "int64",
            "type" : "integer"
          },
          "representation" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "summaryTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tokenBudget" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "CreateMessages" : {
        "properties" : {
          "messages" : {
            "items" : {
              "$ref" : "#/components/schemas/NewMessage"
            },
            "maxItems" : 100,
            "minItems" : 0,
            "type" : "array"
          }
        },
        "required" : [ "messages" ],
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "MessageResponse" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "peer" : {
            "type" : "string"
          },
          "seq" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tokenCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "NewMessage" : {
        "properties" : {
          "content" : {
            "minLength" : 1,
            "type" : "string"
          },
          "metadata" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "peer" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "content", "peer" ],
        "type" : "object"
      },
...
      "PageResponseMessageResponse" : {
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/MessageResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "SearchMessages" : {
        "properties" : {
          "filter" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "A token from `POST /v1/tokens`, in four nesting scopes: admin, workspace, peer, session. Each operation below names the narrowest one that may call it.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Pair-scoped memory. Every fact belongs to a directed (observer, observed) pair: what bob remembers about alice is not what alice remembers about herself, and no route crosses that line.\n\nThree tiers of read: `context` assembles a session's summary and recent messages within a token budget, `recall` ranks conclusions by six fused signals, and `chat` runs an agentic loop over both when the answer needs more than one lookup.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "aimon-memory",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/workspaces/{workspace}/sessions/{session}/context" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.** A session-scoped token may also call this one, but only if it carries `allowMemberRead`.",
        "operationId" : "context",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "session",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "tokens",
          "required" : false,
          "schema" : {
            "default" : 4000,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "target",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "perspective",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Tier 0: summary plus recent messages inside a token budget",
        "tags" : [ "messages" ]
      }
    },
    "/v1/workspaces/{workspace}/sessions/{session}/messages" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.** A session-scoped token may also call this one, but only if it carries `allowMemberRead`.",
        "operationId" : "list",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "session",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseMessageResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "List a session's messages",
        "tags" : [ "messages" ]
      },
      "post" : {
        "description" : "**Requires a `session` token or wider.**",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "session",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "wait",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateMessages"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MessageResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request conflicts with existing data, or names a workspace, peer or session that does not exist. The database refused it, and the reply says no more than that: the driver's message quotes the constraint, the table and often the value."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Store messages and queue derivation",
        "tags" : [ "messages" ]
      }
    },
    "/v1/workspaces/{workspace}/sessions/{session}/messages/search" : {
      "post" : {
        "description" : "**Requires a `peer` token or wider.** A session-scoped token may also call this one, but only if it carries `allowMemberRead`.",
        "operationId" : "search",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "session",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SearchMessages"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseMessageResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The filter parsed but was rejected: an unfilterable column, or an operand of the wrong type. Refused rather than quietly dropped, which would answer 200 with rows nobody asked for."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Search a session's messages",
        "tags" : [ "messages" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "description" : "Relative to wherever this service is deployed.",
    "url" : "/"
  } ],
...
}